requirements.txt
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.0
catboost==1.2.5
streamlit==1.37.0
plotly==5.23.0
matplotlib==3.8.0
seaborn==0.13.2
pyyaml==6.0.2